Move S12Si finding functions to peak_functions
peak_functions.py now contains functions that search for S1, S2 and
S2Si, while pmaps_functions.py contains functions that manipulate
already existing S1, S2 & S2Si objects
